requirements.txt
africastalking==1.2.7
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
contextlib2==21.6.0
cryptography==42.0.7
dj-database-url==2.1.0
dj-rest-auth==6.0.0
Django==5.0.4
django-allauth==0.62.1
django-cors-headers==4.3.1
django-oidc-provider==0.8.2
django-phonenumber-field==7.3.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
future==1.0.0
gunicorn==22.0.0
httplib2==0.22.0
idna==3.7
packaging==24.0
phonenumbers==8.13.36
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodomex==3.20.0
pyjwkest==1.4.2
PyJWT==2.8.0
pyparsing==3.1.2
requests==2.31.0
schema==0.7.5
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
